chore(space-wizard): refactor wizard component into smaller sub-components
