+++ src/app/space-wizard/components/forge-panel/forge-panel.component.html
@@ -1,0 +1,65 @@
+    <header class="container-fluid forge-header">
+      <div class="row">
+        <div class="col-xs-12 wizard-step-title">
+          <h4 class="text" >Welcome to your {{configurator.currentSpace.attributes.name | spaceName}} space.</h4>
+          <div></div>
+          <span (click)="workflow.cancel()" class="pficon pficon-close cursor-pointer"></span>
+        </div>
+      </div>
+      <div class="row" >
+        <div class="col-xs-12 wizard-step-sub-title">
+          <h1 class="text" >How would you like to get started?</h1>
+        </div>
+      </div>
+    </header>
+    <section class="forge-section container-fluid flex-container in-column-direction with-content-stretch padding-standard-offset ">
+      <div class="row flex-grow-1 flex-container">
+        <div class="col-xs-12 panel">
+          <div id="planSpaceButton" class="panel-item cursor-pointer user-select-none">
+            <header>
+              <h3><strong>Plan</strong> out my space</h3>
+            </header>
+            <section><span><br></span></section>
+            <footer><span class="fa fa-3x fa-list" ></span></footer>
+          </div>
+          <div id="newCodeButton" class="panel-item cursor-pointer user-select-none">
+            <header>
+              <h3>Create <strong>new code</strong> from <strong>scratch</strong></h3>
+            </header>
+            <section><span><br></span></section>
+            <footer><span class="fa fa-3x fa-code" ></span></footer>
+          </div>
+          <div id="importCodeButton" class="panel-item cursor-pointer user-select-none" (click)="workflow.gotoNextStep({name:configurator.steps.forgeImportGit})" >
+            <header>
+              <h3>Import existing <strong>code</strong></h3>
+            </header>
+            <section><span><br></span></section>
+            <footer >
+              <span class="pficon fa-3x pficon-import " ></span>
+            </footer>
+          </div>
+          <div id="techStackButton" class="panel-item cursor-pointer user-select-none" (click)="workflow.gotoNextStep({name:configurator.steps.forgeQuickStart})" >
+            <header>
+              <h3>Select <strong>technology stack</strong> and <strong>pipeline</strong></h3>
+            </header>
+            <section><span><br></span></section>
+            <footer class="container-fluid">
+              <div class="row">
+                <div class="col-xs-12">
+                  <span class="pficon fa-3x pficon-screen  " ></span>
+                </div>
+              </div>
+            </footer>
+          </div>
+
+        </div>
+      </div>
+    </section>
+    <footer class="container-fluid padding-standard-offset">
+      <div class="row">
+        <div class="col-xs-12 wizard-step-tool-bar with-items-in-center">
+          <a id="noThanksButton" class="btn btn-primary" (click)="finish()"> No thanks, take me to the
+            {{configurator.currentSpace.attributes.name | spaceName}} dashboard</a>
+        </div>
+      </div>
+    </footer>
