+++ src/app/space-wizard/space-wizard.component.scss
@@ -76,20 +76,6 @@ $wizard-block-header-color: $color-pf-white;
       &.container {
           width:100%;
       }
-      &.forge-header {
-        background-color: $color-pf-black-200;
-        .row {
-          &:first-of-type {
-            background-color: $color-pf-white !important;
-            border-bottom:solid 1px $color-pf-black-300;
-          }
-        }
-        .wizard-step-sub-title {
-          background-color: $color-pf-black-200;
-
-        }
-
-      }
     }
     >section {
       flex:1;
@@ -220,14 +206,6 @@ $wizard-block-header-color: $color-pf-white;
           font-weight: 600;
         }
       }
-      .wizard-step-sub-title {
-        background-color:$color-pf-white;
-        display:flex;
-        justify-content: center;
-        .text {
-          font-weight: 600;
-        }
-      }
       .wizard-step-tool-bar {
         position:relative;
         display:flex;
@@ -254,185 +232,6 @@ $wizard-block-header-color: $color-pf-white;
         &.with-items-stretch{
           justify-content:stretch;
         }
-      }
-    }
-  }
-
-  .wizard-step {
-    section {
-      &.layout-block-container {
-        margin: 0;
-        padding-top: $wizard-spacing;
-        padding-bottom: $wizard-spacing;
-        display: flex;
-        flex-direction: row;
-        flex-wrap: wrap;
-        flex-grow: 1;
-        align-items: flex-start;
-        align-content: flex-start;
-        justify-content: center;
-      }
-      .layout-block {
-        position: relative;
-        display: flex;
-        padding-top: $wizard-spacing;
-        margin: $wizard-spacing;
-        min-width: 150px;
-        border: solid 1px white;
-        box-shadow: $wizard-shadow-offset $wizard-shadow-offset $wizard-shadow-blur-radius $wizard-shadow-spread-radius $wizard-layout-block-shadow-color;
-        &:hover {
-          box-shadow: $wizard-shadow-offset $wizard-shadow-offset $wizard-shadow-blur-radius $wizard-shadow-spread-radius darken($wizard-layout-block-shadow-color, 10%);
-        }
-        &:active {
-          padding-top: $wizard-spacing + 1px;
-          padding-left: 1px;
-          border-bottom: 0;
-          border-right: 0;
-          box-shadow: $wizard-shadow-offset $wizard-shadow-offset $wizard-shadow-blur-radius $wizard-shadow-spread-radius darken($wizard-layout-block-shadow-color, 20%);
-        }
-        .layout-block-icon {
-          margin-left: auto;
-          margin-right: auto;
-        }
-        .layout-block-title {
-          font-size: smaller;
-          font-weight: bolder;
-        }
-        > footer {
-          text-align: center;
-        }
-        > header {
-          text-align: center;
-        }
-        // setting specific to space stack
-        &.layout-block-stack {
-          padding-top: 0;
-          flex-direction: column;
-          flex-grow: 1;
-          > header {
-            border-top: 2px solid $wizard-layout-block-header-border-color;
-            min-height: 2px;
-          }
-        }
-        // settings specific to pipeline selection
-        &.layout-block-pipeline {
-          $stage-padding: 3px;
-          padding-top: 0;
-          flex-direction: column;
-          flex-grow: 1;
-          min-width: 200px;
-          min-height: 170px;
-          > header {
-            background-color: $wizard-block-header-color;
-            padding: $wizard-spacing;
-            border-top: 3px solid $wizard-layout-block-header-border-color;
-          }
-          .stage-block {
-            span.task-block {
-              color: white;
-              background-color: green;
-              border-radius: 1.0rem;
-              padding: $stage-padding 1.0rem;
-              margin-right: $stage-padding;
-              font-size: x-small;
-            }
-          }
-          span.environment-block {
-            color: white;
-            background-color: gray;
-            border-radius: 1.0rem;
-            padding: $stage-padding 1.0rem;
-            margin-right: $stage-padding;
-            font-size: x-small;
-          }
-        }
-      }
-      .panel{
-        display:flex;
-        flex-grow:1;
-        flex-direction: row;
-        justify-content: center;
-        flex-wrap : wrap;
-        background-color: transparent;
-        box-shadow: none;
-
-        .panel-item {
-          // this is a flex container for its children
-          display: flex;
-          //starting width
-          max-width: 180px;
-          flex-basis: 180px;
-          // but it can grow
-          flex-grow: 1;
-          //or shrink
-          flex-shrink:1;
-          // content flex from top to bottom
-          flex-direction: column;
-          //padding: $wizard-spacing;
-          margin: ($grid-gutter-width / 4);
-          //margin-right: $wizard-padding;
-          border:solid 2px transparent;
-          box-shadow: $wizard-shadow-offset  $wizard-shadow-blur-radius $wizard-shadow-spread-radius $wizard-layout-block-shadow-color;
-
-          background-color: $wizard-section-column-background-color;
-          > header {
-            padding-top: $wizard-padding;
-            text-align: center;
-            color: $color-pf-black;
-            font-weight: normal;
-          }
-          > section {
-            text-align: center;
-            flex-grow: 1;
-            padding-top: $wizard-padding;
-          }
-          > div {
-            text-align: center;
-            flex-grow: 1;
-            padding-top: $wizard-padding;
-          }
-          > footer {
-            text-align: center;
-            height: auto;
-            padding-bottom: $wizard-padding;
-          }
-
-          &:hover {
-            border: solid 2px $color-pf-blue-300;
-            box-shadow: $wizard-shadow-offset  $wizard-shadow-blur-radius $wizard-shadow-spread-radius darken($wizard-layout-block-shadow-color, 10%);
-          }
-          &:active {
-            padding-top: $wizard-spacing + 1px;
-            padding-left: $wizard-spacing - 1px;
-            margin-bottom:(($grid-gutter-width / 4) - 1px);
-
-            > header {
-              padding-top: $wizard-padding + 1px;
-              margin-bottom:-1px;
-
-            }
-            > section {
-              padding-top: $wizard-padding + 1px;
-              margin-bottom:-1px;
-
-            }
-            > div {
-              padding-top: $wizard-padding + 1px;
-              margin-bottom:-1px;
-            }
-            > footer {
-              padding-top: 1px;
-              margin-bottom:-1px;
-
-            }
-
-            border: solid 2px $color-pf-blue-400;
-            box-shadow: $wizard-shadow-offset  $wizard-shadow-blur-radius $wizard-shadow-spread-radius darken($wizard-layout-block-shadow-color, 20%);
-          }
-
-
-        }
-
       }
     }
   }
