+++ src/app/space-wizard/space-wizard.module.ts
@@ -8,6 +8,8 @@ import { LoggerFactory } from './common/logger';
 import { ForgeAppGeneratorComponent } from './components/forge-app-generator/forge-app-generator.component';
 import { ForgePipelineViewComponent } from './components/forge-pipeline-view/forge-pipeline-view.component';
 import { ForgeStepViewComponent } from './components/forge-step-view/forge-step-view.component';
+import { ForgePanelComponent } from './components/forge-panel/forge-panel.component';
+import { CreateSpaceComponent } from './components/create-space/create-space.component';
 import { WorkflowFactory } from './models/workflow';
 
 import { IForgeServiceProvider } from './services/forge.service';
@@ -29,6 +31,8 @@ import { CodebasesService } from '../create/codebases/services/codebases.service
     ForgeAppGeneratorComponent,
     ForgePipelineViewComponent,
     ForgeStepViewComponent,
+    ForgePanelComponent,
+    CreateSpaceComponent,
     UniqueSpaceNameValidatorDirective,
     ValidSpaceNameValidatorDirective,
     SelectedItemsPipe,
